requirements.txt
 aiohttp==3.7.2
 Flask==1.1.2
 lxml==4.6.1
 musicbrainzngs==0.7.1
 numpy==1.19.4
 python-dateutil==2.8.1
 pandas==1.1.4
 matplotlib==3.3.3
 gunicorn==20.0.4
